requirements.txt
python-docx==1.1.2
python-dotenv==1.0.1
rich==13.9.4
pdfplumber==0.11.5
pymupdf4llm==0.0.17


playwright==1.49.1
aiohttp==3.11.11
pydantic==2.10.4
pydantic_core==2.27.2
dataclasses_json==0.6.7
nltk==3.9.1
tiktoken==0.8.0
openai==1.59.3
deprecated==1.2.15
tenacity==9.0.0
litellm==1.56.9
pandas==2.2.3
azure_search_documents==11.5.2
nest_asyncio==1.6.0
sqlalchemy==2.0.36
bs4==0.0.2
openpyxl
weaviate
deeplake
weaviate-client
pillow # cause error in deeplake
pypdf==5.1.0